+++ src/bma220/bma220.cxx
@@ -107,9 +107,9 @@ BMA220::BMA220(std::string initStr) : mraaIo(new mraa::MraaIo(initStr))
   for(std::string tok :upmTokens)
   {
     if(tok.substr(0, 9) == "writeReg:") {
-      uint8_t reg = std::stoi(tok.substr(9), &sz, 0);
+      uint8_t reg = std::stoul(tok.substr(9), &sz, 0);
       tok = tok.substr(9);
-      uint8_t val = std::stoi(tok.substr(sz+1), nullptr, 0);
+      uint8_t val = std::stoul(tok.substr(sz+1), nullptr, 0);
       writeReg(reg, val);
     }
     if(tok.substr(0, 22) == "setAccelerometerScale:") {
